# Validation optimizations