--- a/core/src/main/scala/caliban/schema/Types.scala
+++ b/core/src/main/scala/caliban/schema/Types.scala
@@ -160,16 +160,16 @@ object Types {
    */
   def unify(l: List[__Field]): Option[__Type] =
     l.headOption.flatMap { first =>
-      val args                            = first.args.map(_.`type`())
+      val args                            = first.args.map(_._type)
       def _unify(f2: __Field)(t1: __Type) =
         if (
           args.length == f2.args.length &&
-          args.zip(f2.args.map(_.`type`())).forall(v => same(v._1, v._2))
+          args.zip(f2.args.map(_._type)).forall(v => same(v._1, v._2))
         )
-          unify(t1, f2.`type`())
+          unify(t1, f2._type)
         else None
 
-      l.drop(1).foldLeft(Option(first.`type`()))((acc, t) => acc.flatMap(_unify(t)))
+      l.drop(1).foldLeft(Option(first._type))((acc, t) => acc.flatMap(_unify(t)))
     }
 
   /**