--- a/core/src/test/scala/caliban/schema/SchemaSpec.scala
+++ b/core/src/test/scala/caliban/schema/SchemaSpec.scala
@@ -130,7 +130,7 @@ object SchemaSpec extends ZIOSpecDefault {
 
         implicit val somethingSchema: Schema[Any, Something] = Schema.gen[Any, Something].rename("SomethingElse")
 
-        assertTrue(Types.innerType(introspectSubscription[Something]).name.get == "SomethingElse")
+        assertTrue(introspectSubscription[Something].innerType.name.get == "SomethingElse")
       },
       test("union redirect") {
         case class Queries(union: RedirectingUnion)