I was playing around with IntelliJ's new profiler with in-editor hints, and I managed to identify a few places where the code could be optimized. The majority of the optimizations are for validation. 

With these changes, we get between 35-50% performance increase for validation, and about 6% for the fragments benchmark which also includes parsing + execution

Before:

```
[info] Benchmark                                       Mode  Cnt       Score      Error  Units
[info] GraphQLBenchmarks.fragmentsCaliban             thrpt    5     258.842 ±   16.625  ops/s
[info] execution.ValidationBenchmark.deep10000        thrpt    5  119928.824 ± 1033.349  ops/s
[info] execution.ValidationBenchmark.multifield10000  thrpt    5  166221.001 ± 1928.873  ops/s
[info] execution.ValidationBenchmark.simple10000      thrpt    5  270143.275 ± 3360.337  ops/s
```

After:

```
[info] Benchmark                                       Mode  Cnt       Score      Error  Units
[info] GraphQLBenchmarks.fragmentsCaliban             thrpt    5     275.328 ±    3.720  ops/s
[info] execution.ValidationBenchmark.deep10000        thrpt    5  181127.662 ±  915.815  ops/s
[info] execution.ValidationBenchmark.multifield10000  thrpt    5  231313.426 ± 5619.929  ops/s
[info] execution.ValidationBenchmark.simple10000      thrpt    5  360995.697 ± 3407.331  ops/s
```